requirements.txt
asgiref==3.4.1
async-lru==1.0.2
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-smart-selects==1.5.9
django-split-settings==1.1.0
idna==3.3
keyboa==3.1.0
Pillow==8.4.0
pyaes==1.6.1
pyasn1==0.4.8
Pyrogram==1.2.9
PySocks==1.7.1
pyTelegramBotAPI==4.1.1
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
Telethon==1.23.0
TgCrypto==1.2.2
urllib3==1.26.7
